Add entry in Settings.cfg to specify a IP address
Added an Entry to Settings.cfg to specify an IP address for the
PlexConnect DNS Server to be used for intercepting the Trailers Call
from ATV. THe default is 'xxx.xxx.xxx.xxx' and can be changed to any
IPv4 address. This address should match the IP address entered into the
ATV DNS address field.